Fix #15: importing wrong text as mardown.
When Ghost changed to the MobileDoc editor. It also changed where in the
JSON it was storing the Markdown text. Content no longer held it.
PlainText was assumed to now hold the content, but this was wrong. It
was actually being held in a stringified JSON object

This change updates to pull the correct markdown text. Though it should
be revisted as there might be some other edge cases that need to be
considered.