fix #25 missing cards
This fix add the HTML and Bookmark cards to the mobiledoc renderer.
The Bookmark card, is limited, in that it doesn't add all the class
options around the different elements like Ghost does. More work will
need to be done in the upcoming rewrite to allow for class options or
templates for these types.